Update keypadTest.c
You can run this program on its own for testing purposes if you include the timer.h file in the same project folder. I did this purposefully for testing. If it works, I can modify it to be a header file. Since Daniel is using a single file for all of his testing, I thought this would be the easiest format (with the register definitions included) so that he may "throw it in" without using another header file.